The kinetic energy seems to be ok, but the potential energy is large. By looking at the positions of the different particle, it seems like the spin down pariticles (3 and 4) are being moved away from the origin. I think this is what makes the pot. energy so large. Let me check that the equation is correct Hermite.

+++ Hamiltonians/harmonicoscillator.cpp
@@ -27,7 +27,7 @@ double HarmonicOscillator::computePotentialEnergy() {
 
     auto m_particles = m_system->getParticles();
 
-    int numberOfParticles = m_system->getNumberOfParticles();
+    int numberOfParticles = (int) m_system->getNumberOfParticles();
     int numberOfDimensions = m_system->getNumberOfDimensions();
 
     std::vector <double> particlePosition(numberOfDimensions);
